Refactor tests/pyre/PyreApp (move to own file) to allow running app manually when debugging.
